dioxus_history

Struct MemoryHistory

pub struct MemoryHistory { /* private fields */ }

A History provider that stores all navigation information in memory.

Implementations

impl MemoryHistory

pub fn with_initial_path(path: impl ToString) -> Self

Create a MemoryHistory starting at path.

#[derive(Clone, Routable, Debug, PartialEq)]
enum Route {
    #[route("/")]
    Index {},
    #[route("/some-other-page")]
    OtherPage {},
}

let mut history = dioxus_history::MemoryHistory::with_initial_path(Route::Index {});
assert_eq!(history.current_route(), Route::Index {}.to_string());
assert_eq!(history.can_go_back(), false);

pub fn with_prefix(self, prefix: impl ToString) -> Self

Set the base path for the history. All routes will be prefixed with this path when rendered.

let mut history = MemoryHistory::default().with_prefix("/my-app");

// The base path is set to "/my-app"
assert_eq!(history.current_prefix(), Some("/my-app".to_string()));